Smoothed the reverse out + separate worm joint
