+++ src/WormV2.js
@@ -1,11 +1,12 @@
 import { Colors } from "./Colors.js";
 import SnekTongue from "./SnekTongue.js";
 import Vec2 from "./vec2.js";
+import WormJoint from "./WormJoint";
 
-const maxAngle = 0.3;
-const maxDeltaAngle = 0.05;
-const jointStiffness = 0.9;
-const historyLength = 7; //frames of history ish
+export const maxAngle = 0.3;
+export const maxDeltaAngle = 0.05;
+export const jointStiffness = 0.9;
+export const historyLength = 7; //frames of history ish
 const MaxAngularVelocity = 3;
 
 export default class Snek {
@@ -69,21 +70,16 @@ export default class Snek {
 
     }
     
-    // create(p){
-    //     if (this.objects.length < this.partCount) {
-    //         this.createWholeWorm(p);
-    //         return;
-    //     }
-    // }
-
-    move(p) {
-        p = new Vec2(p);
-        this.isMoving = true;
-
+    create(p){
         if (this.objects.length < this.partCount) {
             this.createWholeWorm(p);
             return;
         }
+    }
+
+    move(p) {
+        p = new Vec2(p);
+        this.isMoving = true;
 
         var lastWormElement = this.objects[this.objects.length - 1];
         var lastElemPos = new Vec2(lastWormElement.position);
@@ -131,9 +127,9 @@ export default class Snek {
         var secondToEndToMouse = p.sub(secondLastPos);
         var angle = lastSegmentVec.angle(secondToEndToMouse);
 
+        //This needs to be the other way around for reverse
+        secondLastWormElement.myJoint.setAngle(angle);
         this.shuffleAnglesForwards();
-        lastWormElement.myJoint.setAngle(angle);
-
 
         var relativeTongueAngle = angle;
         var maxAngle = 1;
@@ -159,7 +155,7 @@ export default class Snek {
     }
     shuffleAnglesForwards(){
         //for moving 'backwards'
-        for (let i =  this.objects.length - 2; i >=1; i--) {
+        for (let i = this.objects.length - 2; i >=2; i--) {
             const obj = this.objects[i];
             const nextObj = this.objects[i - 1];
             obj.myJoint.setAngle(nextObj.myJoint.angleHistory[0]);
@@ -205,94 +201,3 @@ export default class Snek {
 
 }
 
-class WormJoint {
-    constructor(a, b, angle, baseLength, matterHander) {
-        this.matterHander = matterHander;
-        this.bodA = a;
-        this.bodB = b;
-        this.baseLength = baseLength;
-        this.angle = angle;
-        this.angleHistory = Array(historyLength).fill(0);
-        this.targetAngle = angle;
-
-
-        //1 = top
-        //p = parallel
-        //d = diagonal
-
-        var a1 = new Vec2(0, baseLength / 2);
-        var a2 = new Vec2(0, -baseLength / 2);
-        var c = new Vec2(0, 0);
-        var f = new Vec2(baseLength / 2, 0);
-
-
-        this.joints = {
-            p1: this.matterHander.createJoint(a, b, a1, a1, jointStiffness),
-            p2: this.matterHander.createJoint(a, b, a2, a2, jointStiffness),
-            d1: this.matterHander.createJoint(a, b, c, a2, jointStiffness),
-            d2: this.matterHander.createJoint(a, b, c, a1, jointStiffness),
-            f: this.matterHander.createJoint(a, b, c, f, jointStiffness),
-        };
-
-        if (!(a.myConstraints)) {
-            a.myConstraints = [];
-        }
-
-
-        for (var i in this.joints) {
-            var joint = this.joints[i];
-            a.myConstraints.push(joint);
-            joint.myStartLength = joint.length + 0;
-            //joint.render.strokeStyle = "#0000";
-        }
-
-
-    }
-
-
-    setColor(angle) {
-        var strain = Math.abs(angle / maxAngle);
-        this.bodA.render.fillStyle = Colors.SnekGradient(strain);
-    }
-
-    setAngle(angle) {
-        if (Math.abs(angle) > maxAngle) {
-            var sign = Math.sign(angle);
-            angle = maxAngle * sign;
-        }
-        this.targetAngle = angle;
-
-
-        if (this.angleHistory.length >= historyLength) {
-            this.angleHistory.shift();
-        }
-        this.angleHistory.push(angle);
-
-
-        var deltaAngle = this.angle - this.targetAngle;
-        if (Math.abs(deltaAngle) < maxDeltaAngle) {
-            this.angle = this.targetAngle;
-        } else {
-            this.angle -= maxDeltaAngle * Math.sign(deltaAngle);
-        }
-
-
-
-        this.setParrallel(this.joints.p1, this.angle);
-        this.setParrallel(this.joints.p2, -this.angle);
-
-        this.setColor(this.angle);
-    }
-
-    setParrallel(j, angle) {
-        var delta = Math.sin(angle) * this.baseLength / 2;
-        var length = j.myStartLength + delta;
-
-        j.length = Math.max(length, 2);
-
-    }
-    update() {
-
-    }
-
-}
